Extract model tests into their own module
