+++ prlint/github/tests/models/test_pull_request.py
@@ -1,0 +1,11 @@
+from django.test import TestCase
+
+from ...models import PullRequest
+
+
+class TestPullRequest(TestCase):
+
+    def test_clean(self):
+        """
+        PullRequest model must be instantiated with 
+        """
